requirements.txt
torch==2.1.0
torchvision==0.16.0
Pillow==10.3.0
numpy==1.26.4
opencv-python==4.8.1.78
basicsr==1.4.2
realesrgan==0.3.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
boto3==1.34.0